+++ src/UI/Fields/Receiver/Receiver.tsx
@@ -1,4 +1,4 @@
-import React, { useState, useEffect, useCallback } from 'react';
+import React, { useState, useEffect, useCallback, useMemo } from 'react';
 import { faExclamationTriangle } from '@fortawesome/free-solid-svg-icons';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { WithClassnameType } from '@multiversx/sdk-dapp/UI/types';
@@ -82,10 +82,16 @@ export const Receiver = (props: WithClassnameType) => {
 
   useEffect(triggerRerenderOnceOnHook, [receiver]);
 
-  const options: GenericOptionType[] = knownAddresses.map((address) => ({
-    value: address,
-    label: address
-  }));
+  const options: GenericOptionType[] = useMemo(
+    () =>
+      knownAddresses
+        ? knownAddresses.map((address) => ({
+            value: address,
+            label: address
+          }))
+        : [],
+    [knownAddresses]
+  );
 
   const onChange = (option: SingleValue<GenericOptionType>) => {
     if (option) {
@@ -120,13 +126,14 @@ export const Receiver = (props: WithClassnameType) => {
         onInputChange={onInputChange}
         inputId='receiverWrapper'
         maxMenuHeight={160}
-        openMenuOnFocus={true}
+        openMenuOnFocus
         isDisabled={getIsDisabled(ValuesEnum.receiver, readonly)}
         options={options}
         filterOption={filterOptions}
         noOptionsMessage={() => null}
         onChange={onChange}
         onBlur={onBlur}
+        isLoading={knownAddresses === null}
         isMulti={false}
         inputValue={inputValue}
         className={classNames(styles.receiverSelectContainer, {
@@ -143,7 +150,8 @@ export const Receiver = (props: WithClassnameType) => {
           Option,
           Placeholder: () => null,
           SingleValue: () => null,
-          IndicatorSeparator: () => null
+          IndicatorSeparator: () => null,
+          LoadingIndicator: () => null
         }}
       />