+++ src/tests/SendESDT/helpers.ts
@@ -1,0 +1,53 @@
+import { fireEvent, RenderResult, queries } from '@testing-library/react';
+import { testAddress, testNetwork } from '__mocks__';
+import { rest, server, mockResponse } from '__mocks__/server';
+import { formConfiguration, renderForm as beginAll } from 'tests/helpers';
+import { ValuesEnum } from 'types';
+
+export { queries };
+
+export const beforAllTokens = (balance?: string) =>
+  beginAll({
+    formConfigValues: {
+      ...formConfiguration,
+      gasLimit: '500000',
+      tokenId: 'TWO-824e70'
+    },
+    ...(balance ? { balance } : {})
+  });
+
+export const twoToken = {
+  identifier: 'TWO-824e70',
+  name: 'TwoTToken',
+  ticker: 'Two',
+  decimals: 2,
+  balance: '100000'
+};
+
+const useInput =
+  (field: ValuesEnum) => (methods: RenderResult) => async (value: string) => {
+    const input: any = await methods.findByTestId(field);
+    const data = { target: { value } };
+    fireEvent.change(input, data);
+    fireEvent.blur(input);
+
+    return input;
+  };
+
+export const useAmountInput = useInput(ValuesEnum.amount);
+export const useGasLimitInput = useInput(ValuesEnum.gasLimit);
+
+export const setupEsdtServer = () => {
+  server.use(
+    rest.get(
+      `${testNetwork.apiAddress}/accounts/${testAddress}/tokens/${twoToken.identifier}`,
+      mockResponse(twoToken)
+    )
+  );
+  server.use(
+    rest.get(
+      `${testNetwork.apiAddress}/accounts/${testAddress}/tokens`,
+      mockResponse([twoToken])
+    )
+  );
+};