+++ src/tests/helpers/TestWrapper.tsx
@@ -5,16 +5,15 @@ import {
   EnvironmentsEnum,
   LoginMethodsEnum
 } from '@multiversx/sdk-dapp/types/enums.types';
+import { GuardianScreenType } from '@multiversx/sdk-dapp/types/transactions.types';
 import { Loader } from '@multiversx/sdk-dapp/UI/Loader';
 import { SendFormContainer, SendFormContainerPropsType } from 'containers';
-import { useSendFormContext } from 'contexts/SendFormProviderContext';
 import {
   useGetInitialValues,
   GetInitialValuesReturnType
 } from 'hooks/useGetInitialValues';
 import getTxType from 'operations/getTxType';
 import { ExtendedValuesType, FormConfigType } from 'types/form';
-import { ConfirmScreen } from 'UI/ConfirmScreen';
 import { Form } from 'UI/Form';
 import { accountConfiguration } from './accountConfiguration';
 import { formConfiguration } from './formConfiguraiton';
@@ -26,6 +25,8 @@ export interface TestWrapperType {
   balance?: string;
   address?: string;
   chainId?: string;
+  GuardianScreen?: (props: GuardianScreenType) => JSX.Element;
+  isGuarded?: boolean;
   ledger?: ExtendedValuesType['ledger'];
 }
 
@@ -34,7 +35,9 @@ export const TestWrapper = ({
   balance = accountConfiguration.balance,
   address = accountConfiguration.address,
   chainId = accountConfiguration.chainId,
-  ledger
+  ledger,
+  isGuarded,
+  GuardianScreen
 }: TestWrapperType) => {
   const initValues = useGetInitialValues({
     configValues: formConfigValues,
@@ -42,6 +45,10 @@ export const TestWrapper = ({
     balance,
     address
   });
+
+  const [hasGuardianScreen, setHasGuardianScreen] = useState(
+    Boolean(GuardianScreen)
+  );
 
   const [isFormSubmitted, setIsFormSubmitted] = useState(
     Boolean(formConfigValues.skipToConfirm)
@@ -83,6 +90,7 @@ export const TestWrapper = ({
     onFormSubmit: () => 'log submit',
     accountInfo: {
       address,
+      isGuarded,
       nonce: accountConfiguration.nonce,
       balance,
       providerType: LoginMethodsEnum.extra
@@ -97,10 +105,8 @@ export const TestWrapper = ({
       setGuardedTransaction: (transaction) => {
         console.log(transaction);
       },
-      setHasGuardianScreen: () => {
-        return false;
-      },
-      hasGuardianScreen: false
+      setHasGuardianScreen,
+      hasGuardianScreen
     },
     tokensInfo: {
       initialNft,
@@ -112,21 +118,10 @@ export const TestWrapper = ({
         })) ?? []
     }
   };
+
   return (
     <SendFormContainer {...containerProps}>
-      <FormContent />
+      <Form GuardianScreen={GuardianScreen} />
     </SendFormContainer>
   );
 };
-
-function FormContent() {
-  const {
-    formInfo: { areValidatedValuesReady }
-  } = useSendFormContext();
-
-  return areValidatedValuesReady ? (
-    <ConfirmScreen providerType={LoginMethodsEnum.extra} />
-  ) : (
-    <Form />
-  );
-}