+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [[0.6.2](https://github.com/multiversx/mx-sdk-dapp-form/pull/178)] - 2023-05-24
+
+- [Loading state support when no known addresses for the Receiver](https://github.com/multiversx/mx-sdk-dapp-form/pull/177)
+- [Added TokenTickerIcon optional param to TokenSelect](https://github.com/multiversx/mx-sdk-dapp-form/pull/176)
+
 ## [[0.6.1](https://github.com/multiversx/mx-sdk-dapp-form/pull/174)] - 2023-05-19
 
 - [Added a new Receiver component with autocomplete support](https://github.com/multiversx/mx-sdk-dapp-form/pull/163)